Start promotion of ReturnType.
