+++ OSBindings/Mac/Clock SignalTests/8088Tests.mm
@@ -19,6 +19,7 @@
 
 #include "NSData+dataWithContentsOfGZippedFile.h"
 
+#include "../../../InstructionSets/x86/AccessType.hpp"
 #include "../../../InstructionSets/x86/Decoder.hpp"
 #include "../../../InstructionSets/x86/Perform.hpp"
 #include "../../../InstructionSets/x86/Flags.hpp"
@@ -97,16 +98,6 @@ struct Memory {
 	public:
 		using AccessType = InstructionSet::x86::AccessType;
 
-		template <typename IntT, AccessType type> struct ReturnType;
-
-		// Reads: return a value directly.
-		template <typename IntT> struct ReturnType<IntT, AccessType::Read> { using type = IntT; };
-		template <typename IntT> struct ReturnType<IntT, AccessType::PreauthorisedRead> { using type = IntT; };
-
-		// Writes: return a reference.
-		template <typename IntT> struct ReturnType<IntT, AccessType::Write> { using type = IntT &; };
-		template <typename IntT> struct ReturnType<IntT, AccessType::ReadModifyWrite> { using type = IntT &; };
-
 		// Constructor.
 		Memory(Registers &registers) : registers_(registers) {
 			memory.resize(1024*1024);
@@ -162,7 +153,7 @@ struct Memory {
 
 		// Accesses an address based on segment:offset.
 		template <typename IntT, AccessType type>
-		typename ReturnType<IntT, type>::type &access(InstructionSet::x86::Source segment, uint16_t offset) {
+		typename InstructionSet::x86::ReturnType<IntT, type>::type &access(InstructionSet::x86::Source segment, uint16_t offset) {
 			if constexpr (std::is_same_v<IntT, uint16_t>) {
 				// If this is a 16-bit access that runs past the end of the segment, it'll wrap back
 				// to the start. So the 16-bit value will need to be a local cache.
@@ -185,7 +176,7 @@ struct Memory {
 
 		// Accesses an address based on physical location.
 		template <typename IntT, AccessType type>
-		typename ReturnType<IntT, type>::type &access(uint32_t address) {
+		typename InstructionSet::x86::ReturnType<IntT, type>::type &access(uint32_t address) {
 			return access<IntT, type>(address, Tag::Accessed);
 		}
